fix binary segmentation handling

+++ learners/losses.py
@@ -55,7 +55,7 @@ class CustomLoss(nn.Module):
             inputs, targets, ignore_index
         )
         return functional.binary_cross_entropy_with_logits(
-            new_inputs, new_targets, weight=weight
+            new_inputs, new_targets.float(), weight=weight
         )
 
     @staticmethod
